refactor(example): split eval_io + alloc out of metal_qwen35_inference main.rs
The teacher-forcing addition pushed main.rs past the 500-LoC cap (539).
Pure code moves per the cap workflow's idiom: the eval-comparison IO
knobs (logits dump, token dump, forced-token list) land in eval_io.rs
and the scratch/state allocation cluster in alloc.rs; main.rs is back
to 456.
